+++ R/helpers.R
@@ -129,7 +129,7 @@ first_run_checks <- function(){
 #' @name CODES_TABLE
 #' @docType data
 #' @author derived from StatCan definitions
-#' @references \url{https://www12.statcan.gc.ca/census-recensement/2016/ref/dict/geo012-eng.cfm}
+#' @references \url{https://www12.statcan.gc.ca/census-recensement/2021/geo/ref/domain-domaine/index2021-eng.cfm?lang=e&id=CSDtype}, \url{https://www12.statcan.gc.ca/census-recensement/2021/geo/ref/domain-domaine/index2021-eng.cfm?lang=e&id=CDtype}
 #' @keywords data
 NULL