+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - add functionality for direct access to StatCan census WDS for 2021
 - add functionality to download original StatCan geographies for 2021
+- update CODES_TABLE for 2021 census
 
 # cancensus 0.5.4